# Code/frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CGDA Backend",
    "description": "Minimal backend service for the geographic analysis platform.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/workflow-runs": {
      "post": {
        "tags": [
          "workflow"
        ],
        "summary": "Submit Workflow",
        "operationId": "submit_workflow_workflow_runs_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowSubmitRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowAcceptedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/workflow-runs/{run_id}": {
      "get": {
        "tags": [
          "workflow"
        ],
        "summary": "Get Workflow Run",
        "operationId": "get_workflow_run_workflow_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowRunStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/workflow-runs/{run_id}/view": {
      "get": {
        "tags": [
          "workflow"
        ],
        "summary": "Get Workflow Run View",
        "operationId": "get_workflow_run_view_workflow_runs__run_id__view_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowRunViewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/workflow-runs/{run_id}/events": {
      "get": {
        "tags": [
          "workflow"
        ],
        "summary": "List Workflow Events",
        "operationId": "list_workflow_events_workflow_runs__run_id__events_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "after_event_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "After Event Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowEventsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/workflow-runs/{run_id}/cancel": {
      "post": {
        "tags": [
          "workflow"
        ],
        "summary": "Cancel Workflow Run",
        "operationId": "cancel_workflow_run_workflow_runs__run_id__cancel_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowRunStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/workflow-runs/{run_id}/retry": {
      "post": {
        "tags": [
          "workflow"
        ],
        "summary": "Retry Workflow Run",
        "operationId": "retry_workflow_run_workflow_runs__run_id__retry_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowAcceptedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AlgorithmWorkflowRequest": {
        "properties": {
          "module_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Module Name"
          },
          "workflow_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workflow Name"
          },
          "workflow_definition": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workflow Definition"
          },
          "workflow_entry_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workflow Entry Name"
          },
          "datasource_selection": {
            "additionalProperties": true,
            "type": "object",
            "title": "Datasource Selection"
          },
          "algorithm_params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Algorithm Params"
          },
          "output_spec": {
            "$ref": "#/components/schemas/AlgorithmOutputSpec"
          },
          "resource_hint": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource Hint"
          },
          "cache_policy": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cache Policy"
          },
          "resume_policy": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resume Policy"
          },
          "tags": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Tags"
          },
          "task_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Type"
          },
          "region": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "time_range": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Range"
          },
          "priority": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          }
        },
        "type": "object",
        "title": "AlgorithmWorkflowRequest"
      },
...
      "ExecutionStatus": {
        "type": "string",
        "enum": [
          "accepted",
          "queued",
          "running",
          "succeeded",
          "failed",
          "cancelled",
          "retry_pending"
        ],
        "title": "ExecutionStatus"
      },
...
      "GeeWorkflowRequest": {
        "properties": {
          "workflow": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workflow"
          },
          "context": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context"
          },
          "workflow_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workflow Id"
          },
          "manifest_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Manifest Uri"
          },
          "update_manifest": {
            "type": "boolean",
            "title": "Update Manifest",
            "default": false
          },
          "tags": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Tags"
          }
        },
        "type": "object",
        "title": "GeeWorkflowRequest",
        "description": "GEE 引擎工作流请求。\n\n与 AlgorithmWorkflowRequest 平行：\n- workflow / context 直接对应 webgis_gee 的 WorkflowSubmissionPayload\n- 当 workflow_definition 非空时，作为 GEE WorkflowDefinition 字典传入\n- manifest_uri 用于导出状态轮询场景（command_type=custom 且仅查询导出状态时使用）"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ResultKind": {
        "type": "string",
        "enum": [
          "json",
          "table",
          "chart",
          "map_layer",
          "log",
          "file",
          "text",
          "diagnostic"
        ],
        "title": "ResultKind"
      },
...
      "SpatialFilter": {
        "properties": {
          "filter_type": {
            "type": "string",
            "title": "Filter Type",
            "default": "bbox"
          },
          "bbox": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BoundingBox"
              },
              {
                "type": "null"
              }
            ]
          },
          "region_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region Code"
          },
          "region_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region Name"
          }
        },
        "type": "object",
        "title": "SpatialFilter"
      },
...
      "TimeRange": {
        "properties": {
          "start_at": {
            "type": "string",
            "format": "date-time",
            "title": "Start At"
          },
          "end_at": {
            "type": "string",
            "format": "date-time",
            "title": "End At"
          },
          "granularity": {
            "$ref": "#/components/schemas/TimeGranularity",
            "default": "hour"
          }
        },
        "type": "object",
        "required": [
          "start_at",
          "end_at"
        ],
        "title": "TimeRange"
      },
...
      "WorkflowAcceptedResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "status": {
            "$ref": "#/components/schemas/ExecutionStatus"
          },
          "status_url": {
            "type": "string",
            "title": "Status Url"
          },
          "events_url": {
            "type": "string",
            "title": "Events Url"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "status",
          "status_url",
          "events_url",
          "created_at",
          "message"
        ],
        "title": "WorkflowAcceptedResponse"
      },
...
      "WorkflowCommandType": {
        "type": "string",
        "enum": [
          "analysis",
          "layer_preview",
          "export",
          "refresh_data",
          "sync_demo",
          "custom"
        ],
        "title": "WorkflowCommandType"
      },
...
      "WorkflowEventsResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/WorkflowEvent"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "items"
        ],
        "title": "WorkflowEventsResponse"
      },
...
      "WorkflowPriority": {
        "type": "string",
        "enum": [
          "low",
          "normal",
          "high",
          "critical"
        ],
        "title": "WorkflowPriority"
      },
...
      "WorkflowResourceProfile": {
        "type": "string",
        "enum": [
          "light",
          "standard",
          "heavy",
          "batch"
        ],
        "title": "WorkflowResourceProfile"
      },
...
      "WorkflowRunStatusResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "status_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Url"
          },
          "events_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Events Url"
          },
          "command_type": {
            "$ref": "#/components/schemas/WorkflowCommandType"
          },
          "command_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Command Label"
          },
          "layer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Layer Id"
          },
          "priority": {
            "$ref": "#/components/schemas/WorkflowPriority",
            "default": "normal"
          },
          "resource_profile": {
            "$ref": "#/components/schemas/WorkflowResourceProfile",
            "default": "standard"
          },
          "realtime_preferred": {
            "type": "boolean",
            "title": "Realtime Preferred",
            "default": false
          },
          "queue_tag": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Queue Tag"
          },
          "status": {
            "$ref": "#/components/schemas/ExecutionStatus"
          },
          "progress": {
            "type": "integer",
            "title": "Progress"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "spatial_filter": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SpatialFilter"
              },
              {
                "type": "null"
              }
            ]
          },
          "time_range": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TimeRange"
              },
              {
                "type": "null"
              }
            ]
          },
          "requested_outputs": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ResultKind"
                },
                {
                  "type": "string"
                }
              ]
            },
            "type": "array",
            "title": "Requested Outputs"
          },
          "client": {
            "$ref": "#/components/schemas/ClientIdentity"
          },
          "map_context": {
            "$ref": "#/components/schemas/RuntimeMapContext"
          },
          "config_overrides": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config Overrides"
          },
          "executor_metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Executor Metadata"
          },
          "result_refs": {
            "items": {
              "$ref": "#/components/schemas/WorkflowResultReference"
            },
            "type": "array",
            "title": "Result Refs"
          },
          "result_dto": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkflowAnalysisResultDto"
              },
              {
                "$ref": "#/components/schemas/WorkflowProviderResultDto"
              },
              {
                "$ref": "#/components/schemas/WorkflowDownloadResultDto"
              },
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Dto"
          },
          "diagnostics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Diagnostics"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "command_type",
          "status",
          "progress",
          "message",
          "created_at",
          "updated_at"
        ],
        "title": "WorkflowRunStatusResponse"
      },
      "WorkflowRunViewResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "subtitle": {
            "type": "string",
            "title": "Subtitle"
          },
          "status_text": {
            "type": "string",
            "title": "Status Text"
          },
          "progress_text": {
            "type": "string",
            "title": "Progress Text"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "metric_rows": {
            "items": {
              "$ref": "#/components/schemas/WorkflowRunViewSummaryRow"
            },
            "type": "array",
            "title": "Metric Rows"
          },
          "result_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Url"
          },
          "artifact_refs": {
            "items": {
              "$ref": "#/components/schemas/WorkflowResultReference"
            },
            "type": "array",
            "title": "Artifact Refs"
          },
          "can_show_link": {
            "type": "boolean",
            "title": "Can Show Link",
            "default": false
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "category",
          "title",
          "subtitle",
          "status_text",
          "progress_text",
          "updated_at"
        ],
        "title": "WorkflowRunViewResponse"
      },
      "WorkflowRunViewSummaryRow": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label"
          },
          "value": {
            "type": "string",
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "label",
          "value"
        ],
        "title": "WorkflowRunViewSummaryRow"
      },
...
      "WorkflowSubmitRequest": {
        "properties": {
          "command_type": {
            "$ref": "#/components/schemas/WorkflowCommandType"
          },
          "command_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Command Label"
          },
          "layer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Layer Id"
          },
          "priority": {
            "$ref": "#/components/schemas/WorkflowPriority",
            "default": "normal"
          },
          "resource_profile": {
            "$ref": "#/components/schemas/WorkflowResourceProfile",
            "default": "standard"
          },
          "realtime_preferred": {
            "type": "boolean",
            "title": "Realtime Preferred",
            "default": false
          },
          "queue_tag": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Queue Tag"
          },
          "spatial_filter": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SpatialFilter"
              },
              {
                "type": "null"
              }
            ]
          },
          "time_range": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TimeRange"
              },
              {
                "type": "null"
              }
            ]
          },
          "parameters": {
            "additionalProperties": true,
            "type": "object",
            "title": "Parameters"
          },
          "algorithm_request": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AlgorithmWorkflowRequest"
              },
              {
                "additionalProperties": true,
                "type": "object"
              }
            ],
            "title": "Algorithm Request"
          },
          "gee_request": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GeeWorkflowRequest"
              },
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gee Request"
          },
          "weather_request": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WeatherWorkflowRequest"
              },
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weather Request"
          },
          "config_overrides": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config Overrides"
          },
          "requested_outputs": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ResultKind"
                },
                {
                  "type": "string"
                }
              ]
            },
            "type": "array",
            "title": "Requested Outputs"
          },
          "client": {
            "$ref": "#/components/schemas/ClientIdentity"
          },
          "map_context": {
            "$ref": "#/components/schemas/RuntimeMapContext"
          },
          "correlation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlation Id"
          },
          "retry_policy": {
            "$ref": "#/components/schemas/RetryPolicy"
          },
          "retry_attempt": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry Attempt"
          }
        },
        "type": "object",
        "required": [
          "command_type"
        ],
        "title": "WorkflowSubmitRequest"
      },
...
    }
  }
}